refactor: 디렉토리 전체 리팩토링
