--- user-api/src/main/java/com/laundreader/userapi/controller/request/chat/ChatStreamRequest.java
@@ -1,10 +1,0 @@
-package com.laundreader.userapi.controller.request.chat;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-
-@Getter
-public class ChatStreamRequest {
-	@NotBlank(message = "빈 문자열 입니다.")
-	private String message;
-}
