--- user-api/src/main/java/com/laundreader/userapi/controller/request/image/ImageValidationRequest.java
@@ -1,31 +1,0 @@
-package com.laundreader.userapi.controller.request.image;
-
-import com.laundreader.userapi.service.dto.ImageDTO;
-import com.laundreader.userapi.service.type.ImageType;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-
-@Getter
-public class ImageValidationRequest {
-	@NotNull(message = "값이 null 일 수 없습니다.")
-	private ImageType type;
-	@NotNull(message = "값이 null 일 수 없습니다.")
-	private Image image;
-
-	public ImageDTO toImageDTO() {
-		return ImageDTO.builder()
-			.format(image.format)
-			.data(image.data)
-			.build();
-	}
-
-	@Getter
-	public static class Image {
-		@NotBlank(message = "빈 문자열 입니다.")
-		private String format;
-		@NotBlank(message = "빈 문자열 입니다.")
-		private String data;
-	}
-}
