+++ user-api/src/main/java/com/laundreader/userapi/dto/image/ImageDTO.java
@@ -1,0 +1,11 @@
+package com.laundreader.userapi.dto.image;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class ImageDTO {
+	String format;
+	String data;
+}
