+++ user-api/src/main/java/com/laundreader/userapi/dto/laundry/HamperDTO.java
@@ -1,0 +1,12 @@
+package com.laundreader.userapi.dto.laundry;
+
+import java.util.List;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class HamperDTO {
+	List<LaundryDTO> laundry;
+}
