+++ user-api/src/main/java/com/laundreader/userapi/dto/laundry/LaundrySymbolDTO.java
@@ -1,0 +1,11 @@
+package com.laundreader.userapi.dto.laundry;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class LaundrySymbolDTO {
+	String code;
+	String description;
+}
