--- user-api/src/main/java/com/laundreader/userapi/service/dto/ImageDTO.java
@@ -1,11 +1,0 @@
-package com.laundreader.userapi.service.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class ImageDTO {
-	String format;
-	String data;
-}
