--- user-api/src/main/java/com/laundreader/userapi/service/dto/LaundryDTO.java
@@ -1,17 +1,0 @@
-package com.laundreader.userapi.service.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Builder
-@Getter
-public class LaundryDTO {
-    List<String> materials;
-    String color;
-    String type;
-    Boolean hasPrintOrTrims;
-    List<String> additionalInfo;
-    List<LaundrySymbolDTO> laundrySymbols;
-}
