--- user-api/src/main/java/com/laundreader/userapi/service/dto/LaundrySymbolDTO.java
@@ -1,11 +1,0 @@
-package com.laundreader.userapi.service.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class LaundrySymbolDTO {
-    String code;
-    String description;
-}
