--- user-api/src/main/java/com/laundreader/userapi/service/response/laundry/HamperSolutionResponseDTO.java
@@ -1,22 +1,0 @@
-package com.laundreader.userapi.service.response.laundry;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-public class HamperSolutionResponseDTO {
-    private List<groupDTO> groups;
-
-    @Getter
-    @AllArgsConstructor
-    public static class groupDTO {
-        private int id;
-        private String name;
-        private String solution;
-        private List<Integer> laundryIds;
-    }
-}
