--- user-api/src/main/java/com/laundreader/userapi/service/response/laundry/SingleSolutionResponseDTO.java
@@ -1,19 +1,0 @@
-package com.laundreader.userapi.service.response.laundry;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-public class SingleSolutionResponseDTO {
-    private List<SolutionDTO> solutions;
-
-    @Getter
-    @AllArgsConstructor
-    public static class SolutionDTO{
-        private String name;
-        private String contents;
-    }
-}
